Add files via upload
This is a Spotify clone website. It has features like play, pause, play next and previous song from the list. The song can also be played by clicking on the play button which is present after the duration of the song list. The information of the song (i.e. song and singer) that the user is playing is also shown at the left bottom corner. Songs and related material in the project are only used to give the project a real music player website look.